data-collection-and-prediction/requirements.txt
Django==4.2.6
django-bootstrap-v5
Pillow
fpdf2
djangorestframework
markdown       
django-filter
django-js-reverse
requests
django-extensions
jupyterlab
gunicorn
django-two-factor-auth
phonenumbers
matplotlib
tqdm
setuptools

numpy
opencv-python
matplotlib
ifaddr
qrcode
django-rosetta
django-mass-edit
celery[redis]
django-celery-results
redis 
imageio
imageio[ffmpeg]
pandas
flower
watchgod
drf_yasg
psycopg2
django-import-export[xlsx]
